+++ src/Support/DomainResolver.php
@@ -16,12 +16,22 @@ class DomainResolver
         return config('ddd.domain_namespace');
     }
 
+    public static function getRelativeObjectNamespace(string $type): string
+    {
+        return config("ddd.namespaces.{$type}", str($type)->plural()->studly()->toString());
+    }
+
+    public static function getDomainObjectNamespace(string $domain, string $type): string
+    {
+        return implode('\\', [static::getConfiguredDomainNamespace(), $domain, static::getRelativeObjectNamespace($type)]);
+    }
+
     public static function guessDomainFromClass(string $class): ?string
     {
         $domainNamespace = Str::finish(DomainResolver::getConfiguredDomainNamespace(), '\\');
 
-        if (! str($class)->startsWith($domainNamespace)) {
-            // Not a domain model
+        if (!str($class)->startsWith($domainNamespace)) {
+            // Not a domain object
             return null;
         }